+++ src/main/kotlin/cacophony/controlflow/generation/CFGGenerator.kt
@@ -14,6 +14,9 @@ import cacophony.semantic.syntaxtree.OperatorUnary
 import cacophony.semantic.syntaxtree.Statement
 import cacophony.semantic.syntaxtree.VariableUse
 
+/**
+ * Converts Expressions into CFG
+ */
 internal class CFGGenerator(
     private val resolvedVariables: ResolvedVariables,
     analyzedUseTypes: UseTypeAnalysisResult,
